Commit Creación BUSXRuta

+++ Client/src/app/bus/ver-bus/ver-bus.component.ts
@@ -4,6 +4,7 @@ import { BusService } from '../shared/bus.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
 import { environment } from '../../../environments/environment';
+import { Conductorxbus } from 'src/app/conductor/shared/conductorxbus';
 
 @Component({
   selector: 'app-ver-bus',
@@ -13,13 +14,15 @@ import { environment } from '../../../environments/environment';
 export class VerBusComponent implements OnInit {
 
   bus: Bus = null;
+  errorMessage = '';
   user = environment.user;
   rol = environment.rol;
+  conductores: Conductorxbus[] = [];
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
-    private busService: BusService
+    private busService: BusService,
   ) { }
 
   ngOnInit(): void {
@@ -31,13 +34,26 @@ export class VerBusComponent implements OnInit {
       console.log(result);
       this.bus = result;
     });
+
+    this.route.paramMap
+    .pipe(
+      switchMap(params => this.busService.getConductoresBus(+params.get('id')))
+    )
+    .subscribe(result => {
+      console.log(result);
+      this.conductores = result;
+    });
   }
 
   eliminarBus(id: number) {
-    this.busService.remove(id).subscribe(
-      resultado => console.log('Bus eliminado!')
-    );
-
+    if (this.conductores.length === 0) {
+      this.busService.remove(id).subscribe(
+        resultado => console.log('Bus eliminado!')
+      );
+      this.router.navigate(['/paginaPrincipalBuses']);
+    } else {
+      this.errorMessage = 'No se puede eliminar al bus. El bus tiene conductores asignados';
+    }
   }
 
   volverLista( ) {
